Rename show filter to toogle code
