+++ lib/livingstyleguide/filters.rb
@@ -36,4 +36,4 @@ require 'livingstyleguide/filters/data'
 require 'livingstyleguide/filters/layout'
 require 'livingstyleguide/filters/syntax'
 require 'livingstyleguide/filters/search_box'
-require 'livingstyleguide/filters/show_hide'
+require 'livingstyleguide/filters/toggle_code'
