+++ lib/livingstyleguide/filters/toggle_code.rb
